refactor: extract trin-execution/e2store + clean dependencies
