+++ ethportal-api/Cargo.toml
@@ -26,7 +26,6 @@ eth_trie = { git = "https://github.com/kolbyml/eth-trie.rs.git", rev = "7e57d3df
 ethereum_serde_utils = "0.5.2"
 ethereum_ssz = "0.5.3"
 ethereum_ssz_derive = "0.5.3"
-ethnum = "1.3.2"
 hex = "0.4.3"
 jsonrpsee = {version="0.20.0", features = ["async-client", "client", "macros", "server"]}
 keccak-hash = "0.10.0"
@@ -42,10 +41,8 @@ secp256k1 = { version = "0.29.0", features = ["global-context", "recovery", "ran
 serde = { version = "1.0.150", features = ["derive", "rc"] }
 serde_json = "1.0.89"
 serde-this-or-that = "0.4.2"
-serde_yaml = "0.9"
 sha2 = "0.10.1"
 sha3 = "0.9.1"
-snap = "1.1.0"
 superstruct = "0.7.0"
 ssz_types = { git = "https://github.com/KolbyML/ssz_types.git", rev = "2a5922de75f00746890bf4ea9ad663c9d5d58efe" }
 thiserror = "1.0.57"
@@ -61,6 +58,8 @@ validator = { version = "0.13.0", features = ["derive"] }
 env_logger = "0.9.0"
 quickcheck = "1.0.3"
 rstest = "0.18.2"
+serde_yaml = "0.9"
+snap = "1.1.0"
 test-log = { version = "0.2.11", features = ["trace"] }
 tracing = "0.1.36"
 tracing-subscriber = "0.3.15"
