+++ portal-bridge/src/bridge/era1.rs
@@ -6,9 +6,12 @@ use std::{
 
 use alloy_primitives::B256;
 use anyhow::{anyhow, ensure};
+use e2store::{
+    era1::{BlockTuple, Era1},
+    utils::get_shuffled_era1_files,
+};
 use futures::future::join_all;
 use rand::{seq::SliceRandom, thread_rng};
-use scraper::{Html, Selector};
 use surf::{Client, Config};
 use tokio::{
     sync::{OwnedSemaphorePermit, Semaphore},
@@ -26,10 +29,7 @@ use crate::{
     },
     gossip::gossip_history_content,
     stats::{HistoryBlockStats, StatsReporter},
-    types::{
-        era1::{BlockTuple, Era1},
-        mode::{BridgeMode, FourFoursMode},
-    },
+    types::mode::{BridgeMode, FourFoursMode},
 };
 use ethportal_api::{
     jsonrpsee::http_client::HttpClient,
@@ -40,9 +40,6 @@ use ethportal_api::{
 use trin_validation::{
     constants::EPOCH_SIZE, header_validator::HeaderValidator, oracle::HeaderOracle,
 };
-
-const ERA1_DIR_URL: &str = "https://era1.ethportal.net/";
-const ERA1_FILE_COUNT: usize = 1897;
 
 pub struct Era1Bridge {
     pub mode: BridgeMode,
@@ -578,37 +575,6 @@ impl Era1Bridge {
         let content_value = HistoryContentValue::Receipts(block_tuple.receipts.receipts);
         gossip_history_content(portal_client, content_key, content_value, block_stats).await
     }
-}
-
-/// Fetches era1 files hosted on era1.ethportal.net and shuffles them
-pub async fn get_shuffled_era1_files(http_client: &Client) -> anyhow::Result<Vec<String>> {
-    let index_html = http_client
-        .get(ERA1_DIR_URL)
-        .recv_string()
-        .await
-        .map_err(|e| anyhow!("{e}"))?;
-    let index_html = Html::parse_document(&index_html);
-    let selector = Selector::parse("a[href*='mainnet-']").expect("to be able to parse selector");
-    let mut era1_files: Vec<String> = index_html
-        .select(&selector)
-        .map(|element| {
-            let href = element
-                .value()
-                .attr("href")
-                .expect("to be able to get href");
-            format!("{ERA1_DIR_URL}{href}")
-        })
-        .collect();
-    ensure!(
-        era1_files.len() == ERA1_FILE_COUNT,
-        format!(
-            "invalid era1 source, not enough era1 files found: expected {}, found {}",
-            ERA1_FILE_COUNT,
-            era1_files.len()
-        )
-    );
-    era1_files.shuffle(&mut thread_rng());
-    Ok(era1_files)
 }
 
 fn get_epoch_from_era1_path(era1_path: &str) -> anyhow::Result<u64> {
