+++ portalnet/Cargo.toml
@@ -12,16 +12,13 @@ authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 
 [dependencies]
 alloy-primitives = "0.7.0"
-alloy-rlp = "0.3.4"
 anyhow = "1.0.68"
 async-trait = "0.1.64"
-base64 = "0.13.0"
 bytes = "1.3.0"
 delay_map = "0.3.0"
 directories = "3.0"
 discv5 = { version = "0.4.1", features = ["serde"] }
 ethereum_ssz = "0.5.3"
-ethereum_ssz_derive = "0.5.3"
 ethportal-api = { path = "../ethportal-api" }
 fnv = "1.0.7"
 futures = "0.3.21"
@@ -35,7 +32,6 @@ lru = "0.7.8"
 parking_lot = "0.11.2"
 rand = "0.8.4"
 serde = { version = "1.0.150", features = ["derive"] }
-serde_json = "1.0.89"
 smallvec = "1.8.0"
 ssz_types = { git = "https://github.com/KolbyML/ssz_types.git", rev = "2a5922de75f00746890bf4ea9ad663c9d5d58efe" }
 stunclient = "0.1.2"
@@ -48,9 +44,7 @@ trin-metrics = { path = "../trin-metrics" }
 trin-storage = { path = "../trin-storage" }
 trin-utils = { path = "../trin-utils" }
 trin-validation = { path = "../trin-validation" }
-url = "2.3.1"
 utp-rs = { git = "https://github.com/ethereum/utp", tag = "v0.1.0-alpha.12" }
-validator = { version = "0.13.0", features = ["derive"] }
 
 [target.'cfg(windows)'.dependencies]
 uds_windows = "1.0.1"
