+++ trin-beacon/Cargo.toml
@@ -29,7 +29,6 @@ tracing = "0.1.36"
 trin-metrics = { path = "../trin-metrics" }
 trin-storage = { path = "../trin-storage" }
 trin-validation = { path = "../trin-validation" }
-trin-utils = { path = "../trin-utils" }
 utp-rs = { git = "https://github.com/ethereum/utp", tag = "v0.1.0-alpha.12" }
 
 [dev-dependencies]
