+++ trin-execution/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "trin-execution"
+version = "0.1.0"
+edition = "2021"
+repository = "https://github.com/ethereum/trin/tree/master/trin-execution"
+license = "GPL-3.0"
+readme = "README.md"
+keywords = ["ethereum", "execution-layer"]
+categories = ["cryptography::cryptocurrencies"]
+description = "Trin's execution used for gossiping state and soon an execution layer client for Ethereum?"
+authors = ["https://github.com/ethereum/trin/graphs/contributors"]
+
+
+[dependencies]
+alloy-consensus = { git = "https://github.com/alloy-rs/alloy.git", rev = "d68a6b787b2904061f0ae7fcc02ece8513e3c500"}
+alloy-primitives = "0.7.0"
+alloy-rlp = "0.3.4"
+anyhow = "1.0.68"
+directories = "3.0"
+ethportal-api = { path = "../ethportal-api" }
+e2store = { path = "../e2store" }
+eth_trie = { git = "https://github.com/kolbyml/eth-trie.rs.git", rev = "7e57d3dfadee126cc9fda2696fb039bf7b6ed688" }
+hashbrown = "0.14.0"
+parking_lot = "0.11.2"
+revm = { version = "8.0.0", features = ["std", "secp256k1", "serde-json"], default-features = false }
+revm-primitives = { version = "3.1.0", features = ["std", "serde"], default-features = false }
+rocksdb = "0.22.0"
+serde = { version = "1.0.150", features = ["derive", "rc"] }
+serde_json = "1.0.89"
+tempfile = "3.3.0"
+thiserror = "1.0.57"
+tracing = "0.1.36"
+
+[dev-dependencies]
+surf = { version = "2.3.2", default-features = false, features = ["h1-client-rustls", "middleware-logger", "encoding"] } # we use rustls because OpenSSL cause issues compiling on aarch64
+test-log = { version = "0.2.11", features = ["trace"] }
+tokio = { version = "1.14.0", features = ["full"] }
+tracing-test = "0.1"
