+++ trin-execution/src/storage/utils.rs
