+++ trin-history/Cargo.toml
@@ -18,15 +18,11 @@ ethereum_ssz = "0.5.3"
 ethportal-api = {path = "../ethportal-api"}
 parking_lot = "0.11.2"
 portalnet = { path = "../portalnet" }
-r2d2 = "0.8.9"
-r2d2_sqlite = "0.24.0"
 serde_json = "1.0.89"
 tokio = { version = "1.14.0", features = ["full"] }
 tracing = "0.1.36"
 tree_hash = { git = "https://github.com/KolbyML/tree_hash.git", rev = "8aaf8bb4184148768d48e2cfbbdd0b95d1da8730" }
-trin-metrics = { path = "../trin-metrics" }
 trin-storage = { path = "../trin-storage" }
-trin-utils = { path = "../trin-utils" }
 trin-validation = { path = "../trin-validation" }
 utp-rs = { git = "https://github.com/ethereum/utp", tag = "v0.1.0-alpha.12" }
