+++ trin-state/Cargo.toml
@@ -20,7 +20,6 @@ ethportal-api = { path = "../ethportal-api" }
 keccak-hash = "0.10.0"
 parking_lot = "0.11.2"
 portalnet = { path = "../portalnet" }
-serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.89"
 thiserror = "1.0.57"
 tokio = {version = "1.14.0", features = ["full"]}
@@ -32,6 +31,7 @@ utp-rs = { git = "https://github.com/ethereum/utp", tag = "v0.1.0-alpha.12" }
 [dev-dependencies]
 env_logger = "0.9.0"
 rstest = "0.18.2"
+serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
 serial_test = "0.5.1"
 test-log = { version = "0.2.11", features = ["trace"] }
